change topicPartition to partition in ClusterInfo.proto. move code to ByteUtils.
