+++ common/src/main/java/org/astraea/common/consumer/Deserializer.java
@@ -16,17 +16,8 @@
  */
 package org.astraea.common.consumer;
 
-import java.nio.ByteBuffer;
-import java.util.Arrays;
 import java.util.Base64;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import org.apache.kafka.common.header.Headers;
 import org.apache.kafka.common.serialization.ByteArrayDeserializer;
 import org.apache.kafka.common.serialization.DoubleDeserializer;
@@ -36,15 +27,7 @@ import org.apache.kafka.common.serialization.LongDeserializer;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.astraea.common.ByteUtils;
 import org.astraea.common.Header;
-import org.astraea.common.Utils;
 import org.astraea.common.admin.ClusterInfo;
-import org.astraea.common.admin.Config;
-import org.astraea.common.admin.NodeInfo;
-import org.astraea.common.admin.Replica;
-import org.astraea.common.admin.Topic;
-import org.astraea.common.admin.TopicPartition;
-import org.astraea.common.backup.ByteUtils;
-import org.astraea.common.generated.BeanObjectOuterClass;
 import org.astraea.common.json.JsonConverter;
 import org.astraea.common.json.TypeRef;
 import org.astraea.common.metrics.BeanObject;
@@ -93,10 +76,8 @@ public interface Deserializer<T> {
   Deserializer<Long> LONG = of(new LongDeserializer());
   Deserializer<Float> FLOAT = of(new FloatDeserializer());
   Deserializer<Double> DOUBLE = of(new DoubleDeserializer());
-  Deserializer<NodeInfo> NODE_INFO = new NodeInfoDeserializer();
-  Deserializer<Topic> TOPIC = new TopicDeserializer();
-  Deserializer<Replica> REPLICA = new ReplicaDeserializer();
-  Deserializer<ClusterInfo> CLUSTER_INFO = new ClusterInfoDeserializer();
+  Deserializer<ClusterInfo> CLUSTER_INFO =
+      (topic, headers, data) -> ByteUtils.readClusterInfo(data);
   Deserializer<BeanObject> BEAN_OBJECT = (topic, headers, data) -> ByteUtils.readBeanObject(data);
 
   /**
@@ -124,68 +105,6 @@ public interface Deserializer<T> {
       else {
         return jackson.fromJson(Deserializer.STRING.deserialize(topic, headers, data), typeRef);
       }
-    }
-  }
-
-  class ClusterInfoDeserializer implements Deserializer<NodeInfo> {
-    @Override
-    public ClusterInfo deserialize(String topic, List<Header> headers, byte[] data) {
-      var outerClusterInfo =
-          Utils.packException(() -> ClusterInfoOuterClass.ClusterInfo.parseFrom(data));
-      return ClusterInfo.of(
-          outerClusterInfo.getClusterId(),
-          outerClusterInfo.getNodeInfoList().stream()
-              .map(
-                  nodeInfo -> NodeInfo.of(nodeInfo.getId(), nodeInfo.getHost(), nodeInfo.getPort()))
-              .collect(Collectors.toList()),
-          outerClusterInfo.getTopicList().stream()
-              .map(
-                  protoTopic ->
-                      new Topic() {
-                        @Override
-                        public String name() {
-                          return protoTopic.getName();
-                        }
-
-                        @Override
-                        public Config config() {
-                          return Config.of(protoTopic.getConfigMap());
-                        }
-
-                        @Override
-                        public boolean internal() {
-                          return protoTopic.getInternal();
-                        }
-
-                        @Override
-                        public Set<TopicPartition> topicPartitions() {
-                          return protoTopic.getTopicPartitionList().stream()
-                              .map(tp -> TopicPartition.of(protoTopic.getName(), tp))
-                              .collect(Collectors.toSet());
-                        }
-                      })
-              .collect(Collectors.toMap(Topic::name, Function.identity())),
-          outerClusterInfo.getReplicaList().stream()
-              .map(
-                  replica ->
-                      Replica.builder()
-                          .topic(replica.getTopic())
-                          .partition(replica.getPartition())
-                          .nodeInfo(
-                              NodeInfo.of(
-                                  replica.getNodeInfo().getId(),
-                                  replica.getNodeInfo().getHost(),
-                                  replica.getNodeInfo().getPort()))
-                          .lag(replica.getLag())
-                          .size(replica.getSize())
-                          .isLeader(replica.getIsLeader())
-                          .isSync(replica.getIsSync())
-                          .isFuture(replica.getIsFuture())
-                          .isOffline(replica.getIsOffline())
-                          .isPreferredLeader(replica.getIsPreferredLeader())
-                          .path(replica.getPath())
-                          .build())
-              .collect(Collectors.toList()));
     }
   }
 }
